# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_start_overlay_translate_api_overlay_pdf_start_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_start_overlay_translate_api_overlay_pdf_start_post",
        "type": "object"
      },
      "Body_start_pdf_translate_api_translate_pdf_start_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "target_lang": {
            "default": "zh",
            "title": "Target Lang",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_start_pdf_translate_api_translate_pdf_start_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartTaskResponse": {
        "properties": {
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "StartTaskResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PDF 阅读翻译后端",
    "version": "0.2.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/edit/pdf/result/{edit_id}": {
      "get": {
        "description": "下载编辑后的新 PDF。",
        "operationId": "result_api_edit_pdf_result__edit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "title": "Edit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Result",
        "tags": [
          "pdf-edit"
        ]
      }
    },
...
    "/api/overlay/pdf/progress/{task_id}": {
      "get": {
        "description": "SSE 进度流：每连接独立订阅，多并发连接各自收到全量事件（审计 3.2）。",
        "operationId": "overlay_progress_api_overlay_pdf_progress__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overlay Progress",
        "tags": [
          "overlay-translate"
        ]
      }
    },
    "/api/overlay/pdf/result/{task_id}": {
      "get": {
        "description": "下载覆盖翻译后的 PDF。",
        "operationId": "overlay_result_api_overlay_pdf_result__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overlay Result",
        "tags": [
          "overlay-translate"
        ]
      }
    },
    "/api/overlay/pdf/start": {
      "post": {
        "description": "接收 PDF，创建覆盖翻译任务（跳过 pdf2zh），立即返回 task_id。",
        "operationId": "start_overlay_translate_api_overlay_pdf_start_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-llm-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Llm-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-llm-base-url",
            "required": false,
            "schema": {
              "default": "https://api.openai.com/v1",
              "title": "X-Llm-Base-Url",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-llm-model",
            "required": false,
            "schema": {
              "default": "gpt-4o-mini",
              "title": "X-Llm-Model",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-llm-max-concurrency",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 8,
              "minimum": 1,
              "title": "X-Llm-Max-Concurrency",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-llm-request-interval-ms",
            "required": false,
            "schema": {
              "default": 1100,
              "maximum": 60000,
              "minimum": 0,
              "title": "X-Llm-Request-Interval-Ms",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-llm-max-retries",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 10,
              "minimum": 0,
              "title": "X-Llm-Max-Retries",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-llm-retry-base-seconds",
            "required": false,
            "schema": {
              "default": 2.0,
              "maximum": 60.0,
              "minimum": 0.1,
              "title": "X-Llm-Retry-Base-Seconds",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_start_overlay_translate_api_overlay_pdf_start_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Overlay Translate",
        "tags": [
          "overlay-translate"
        ]
      }
    },
...
    "/api/translate/pdf/progress/{task_id}": {
      "get": {
        "description": "SSE 进度流：每连接独立订阅，多并发连接各自收到全量事件（审计 3.2）。",
        "operationId": "pdf_progress_api_translate_pdf_progress__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pdf Progress",
        "tags": [
          "pdf-translate"
        ]
      }
    },
    "/api/translate/pdf/result/{task_id}": {
      "get": {
        "description": "下载翻译后的 PDF。",
        "operationId": "pdf_result_api_translate_pdf_result__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pdf Result",
        "tags": [
          "pdf-translate"
        ]
      }
    },
    "/api/translate/pdf/start": {
      "post": {
        "description": "接收 PDF，创建翻译任务，立即返回 task_id。",
        "operationId": "start_pdf_translate_api_translate_pdf_start_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-llm-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Llm-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-llm-base-url",
            "required": false,
            "schema": {
              "default": "https://api.openai.com/v1",
              "title": "X-Llm-Base-Url",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-llm-model",
            "required": false,
            "schema": {
              "default": "gpt-4o-mini",
              "title": "X-Llm-Model",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-llm-max-concurrency",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 8,
              "minimum": 1,
              "title": "X-Llm-Max-Concurrency",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-llm-request-interval-ms",
            "required": false,
            "schema": {
              "default": 1100,
              "maximum": 60000,
              "minimum": 0,
              "title": "X-Llm-Request-Interval-Ms",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-llm-max-retries",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 10,
              "minimum": 0,
              "title": "X-Llm-Max-Retries",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-llm-retry-base-seconds",
            "required": false,
            "schema": {
              "default": 2.0,
              "maximum": 60.0,
              "minimum": 0.1,
              "title": "X-Llm-Retry-Base-Seconds",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_start_pdf_translate_api_translate_pdf_start_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Pdf Translate",
        "tags": [
          "pdf-translate"
        ]
      }
    },
...
  }
}